+++ api/vpc/v1alpha2/vpc_types.go
@@ -89,7 +89,7 @@ type VPCDHCP struct {
 	// InterfaceMTU (optional) is the MTU setting that the dhcp server will send to the clients. It is dependent on the client to honor this option.
 	// +kubebuilder:validation:Minimum: 96
 	// +kubebuilder:validation:Maximum: 9036
-	InterfaceMTU uint32 `json:"interfaceMTU"`
+	InterfaceMTU uint16 `json:"interfaceMTU"`
 }
 
 // VPCDHCPRange defines the DHCP range for the subnet if DHCP server is enabled
